+++ booster/library/Booster/JsonRpc.hs
@@ -23,10 +23,8 @@ import Control.Monad.IO.Class
 import Control.Monad.Trans.Except (catchE, except, runExcept, runExceptT, throwE, withExceptT)
 import Crypto.Hash (SHA256 (..), hashWith)
 import Data.Bifunctor (second)
-import Data.Coerce (coerce)
 import Data.Foldable
 import Data.List (singleton)
-import Data.List.NonEmpty qualified as NonEmpty
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Maybe (catMaybes, fromMaybe, isJust, mapMaybe)
@@ -38,7 +36,6 @@ import Data.Text qualified as Text
 import Data.Text.Encoding qualified as Text
 import GHC.Records
 import Numeric.Natural
-import Prettyprinter (comma, hsep, punctuate, (<+>))
 
 import Booster.CLOptions (RewriteOptions (..))
 import Booster.Definition.Attributes.Base (UniqueId, getUniqueId, uniqueId)
@@ -49,8 +46,6 @@ import Booster.Log
 import Booster.Pattern.ApplyEquations qualified as ApplyEquations
 import Booster.Pattern.Base (Pattern (..), Sort (SortApp), Term, Variable)
 import Booster.Pattern.Base qualified as Pattern
-import Booster.Pattern.Bool (pattern TrueBool)
-import Booster.Pattern.Match (FailReason (..), MatchResult (..), MatchType (..), matchTerms)
 import Booster.Pattern.Pretty
 import Booster.Pattern.Rewrite (
     RewriteConfig (..),
@@ -65,16 +60,14 @@ import Booster.Pattern.Util (
     substituteInPredicate,
     substituteInTerm,
  )
-import Booster.Prettyprinter (renderDefault, renderText)
+import Booster.Prettyprinter (renderText)
 import Booster.SMT.Interface qualified as SMT
 import Booster.Syntax.Json (KoreJson (..), addHeader, prettyPattern, sortOfJson)
 import Booster.Syntax.Json.Externalise
 import Booster.Syntax.Json.Internalise (
     InternalisedPredicates (..),
-    PatternOrTopOrBottom (..),
     TermOrPredicates (..),
     internalisePattern,
-    internalisePatternOrTopOrBottom,
     internaliseTermOrPredicate,
     logPatternError,
     patternErrorToRpcError,
@@ -87,7 +80,6 @@ import Booster.Syntax.ParsedKore.Base qualified as ParsedModule (ParsedModule (.
 import Booster.Syntax.ParsedKore.Internalise (
     addToDefinitions,
     definitionErrorToRpcError,
-    extractExistentials,
  )
 import Booster.Util (Flag (..), constructorName)
 import Kore.JsonRpc.Error qualified as RpcError
@@ -97,6 +89,7 @@ import Kore.JsonRpc.Types.Log
 import Kore.Syntax.Json.Types (Id (..))
 import Kore.Syntax.Json.Types qualified as KoreJson
 import Kore.Syntax.Json.Types qualified as Syntax
+import Booster.Pattern.Implies (runImplies)
 
 respond ::
     forall m.
@@ -419,154 +412,7 @@ respond stateVar request =
                                             { satisfiable = RpcTypes.Unknown
                                             , substitution = Nothing
                                             }
-            RpcTypes.Implies req -> withModule req._module $ \(def, mLlvmLibrary, mSMTOptions, _) -> Booster.Log.withContext CtxImplies $ do
-                solver <- maybe (SMT.noSolver) (SMT.initSolver def) mSMTOptions
-                -- internalise given constrained term
-                let internalised korePat' =
-                        let (korePat, existentials) = extractExistentials korePat'
-                         in runExcept $
-                                internalisePatternOrTopOrBottom DisallowAlias CheckSubsorts Nothing def existentials korePat
-
-                    checkImplies patL substitutionL unsupportedL existsL patR substitutionR unsupportedR existsR = do
-                        let freeVarsL =
-                                ( freeVariables patL.term
-                                    <> (Set.unions $ Set.map (freeVariables . coerce) patL.constraints)
-                                    <> (Set.fromList $ Map.keys substitutionL)
-                                )
-                                    Set.\\ Set.fromList existsL
-                            freeVarsR =
-                                ( freeVariables patR.term
-                                    <> (Set.unions $ Set.map (freeVariables . coerce) patR.constraints)
-                                    <> (Set.fromList $ Map.keys substitutionR)
-                                )
-                                    Set.\\ Set.fromList existsR
-                            freeVarsRminusL = freeVarsR Set.\\ freeVarsL
-                        if (not $ null freeVarsRminusL)
-                            then
-                                pure . Left . RpcError.backendError . RpcError.ImplicationCheckError $
-                                    RpcError.ErrorWithContext "The RHS must not have free variables not present in the LHS" $
-                                        map (pack . renderDefault . pretty' @mods) $
-                                            Set.toList freeVarsRminusL
-                            else do
-                                unless (null unsupportedL && null unsupportedR) $ do
-                                    logMessage'
-                                        ("aborting due to unsupported predicate parts" :: Text)
-                                    unless (null unsupportedL) $
-                                        withContext CtxDetail $
-                                            logMessage
-                                                (Text.unwords $ map prettyPattern unsupportedL)
-                                    unless (null unsupportedR) $
-                                        withContext CtxDetail $
-                                            logMessage
-                                                (Text.unwords $ map prettyPattern unsupportedR)
-                                let
-                                    -- apply the given substitution before doing anything else
-                                    substPatL =
-                                        Pattern
-                                            { term = substituteInTerm substitutionL patL.term
-                                            , constraints = Set.map (substituteInPredicate substitutionL) patL.constraints
-                                            , ceilConditions = patL.ceilConditions
-                                            }
-                                    substPatR =
-                                        Pattern
-                                            { term = substituteInTerm substitutionR patR.term
-                                            , constraints = Set.map (substituteInPredicate substitutionR) patR.constraints
-                                            , ceilConditions = patR.ceilConditions
-                                            }
-
-                                SMT.isSat solver (Set.toList substPatL.constraints) >>= \case
-                                    SMT.IsUnsat ->
-                                        let sort = externaliseSort $ sortOfPattern substPatL
-                                         in implies' (Syntax.KJBottom sort) sort req.antecedent.term req.consequent.term mempty
-                                    _ -> checkImpliesMatchTerms existsL substPatL existsR substPatR
-
-                    checkImpliesMatchTerms existsL substPatL existsR substPatR =
-                        case matchTerms Booster.Pattern.Match.Implies def substPatR.term substPatL.term of
-                            MatchFailed (SubsortingError sortError) ->
-                                pure . Left . RpcError.backendError . RpcError.ImplicationCheckError . RpcError.ErrorOnly . pack $
-                                    show sortError
-                            MatchFailed{} ->
-                                doesNotImply (sortOfPattern substPatL) req.antecedent.term req.consequent.term
-                            MatchIndeterminate remainder ->
-                                ApplyEquations.evaluatePattern def mLlvmLibrary solver mempty substPatL >>= \case
-                                    (Right simplifedSubstPatL, _) ->
-                                        if substPatL == simplifedSubstPatL
-                                            then
-                                                pure . Left . RpcError.backendError . RpcError.ImplicationCheckError . RpcError.ErrorOnly . pack $
-                                                    "match remainder: "
-                                                        <> renderDefault
-                                                            ( hsep $
-                                                                punctuate comma $
-                                                                    map (\(t1, t2) -> pretty' @mods t1 <+> "==" <+> pretty' @mods t2) $
-                                                                        NonEmpty.toList remainder
-                                                            )
-                                            else checkImpliesMatchTerms existsL simplifedSubstPatL existsR substPatR
-                                    (Left err, _) ->
-                                        pure . Left . RpcError.backendError $ RpcError.Aborted (Text.pack . constructorName $ err)
-                            MatchSuccess subst -> do
-                                let filteredConsequentPreds =
-                                        Set.map (substituteInPredicate subst) substPatR.constraints `Set.difference` substPatL.constraints
-
-                                if null filteredConsequentPreds
-                                    then
-                                        implies
-                                            (sortOfPattern substPatL)
-                                            (externaliseExistTerm existsL substPatL.term)
-                                            (externaliseExistTerm existsR substPatR.term)
-                                            subst
-                                    else
-                                        ApplyEquations.evaluateConstraints def mLlvmLibrary solver mempty filteredConsequentPreds >>= \case
-                                            (Right newPreds, _) ->
-                                                if all (== Pattern.Predicate TrueBool) newPreds
-                                                    then
-                                                        implies
-                                                            (sortOfPattern substPatL)
-                                                            (externaliseExistTerm existsL substPatL.term)
-                                                            (externaliseExistTerm existsR substPatR.term)
-                                                            subst
-                                                    else pure . Left . RpcError.backendError $ RpcError.Aborted "unknown constraints"
-                                            (Left other, _) ->
-                                                pure . Left . RpcError.backendError $ RpcError.Aborted (Text.pack . constructorName $ other)
-
-                case (internalised req.antecedent.term, internalised req.consequent.term) of
-                    (Left patternError, _) -> do
-                        void $ Booster.Log.withContext CtxInternalise $ logPatternError patternError
-                        pure $
-                            Left $
-                                RpcError.backendError $
-                                    RpcError.CouldNotVerifyPattern
-                                        [ patternErrorToRpcError patternError
-                                        ]
-                    (_, Left patternError) -> do
-                        void $ Booster.Log.withContext CtxInternalise $ logPatternError patternError
-                        pure $
-                            Left $
-                                RpcError.backendError $
-                                    RpcError.CouldNotVerifyPattern
-                                        [ patternErrorToRpcError patternError
-                                        ]
-                    (Right (IsBottom sort), Right _) ->
-                        implies' (Syntax.KJBottom sort) sort req.antecedent.term req.consequent.term mempty
-                    (Right IsTop{}, _) ->
-                        pure . Left . RpcError.backendError . RpcError.ImplicationCheckError . RpcError.ErrorOnly $
-                            "The check implication step expects the antecedent term to be function-like."
-                    ( Right (IsPattern (existsL, (patL, substitutionL, unsupportedL)))
-                        , Right (IsPattern (existsR, (patR, substitutionR, unsupportedR)))
-                        ) ->
-                            checkImplies patL substitutionL unsupportedL existsL patR substitutionR unsupportedR existsR
-                    (Right IsPattern{}, Right (IsTop sort)) ->
-                        implies' (Syntax.KJTop sort) sort req.antecedent.term req.consequent.term mempty
-                    (Right IsPattern{}, Right (IsBottom sort)) ->
-                        doesNotImply'
-                            sort
-                            ( Just $
-                                RpcTypes.Condition
-                                    { predicate = addHeader $ Syntax.KJBottom sort
-                                    , substitution = addHeader $ Syntax.KJTop sort
-                                    }
-                            )
-                            req.antecedent.term
-                            req.consequent.term
+            RpcTypes.Implies req -> withModule req._module $ \(def, mLlvmLibrary, mSMTOptions, _) -> runImplies def mLlvmLibrary mSMTOptions req.antecedent req.consequent
 
             -- this case is only reachable if the cancel appeared as part of a batch request
             RpcTypes.Cancel -> pure $ Left RpcError.cancelUnsupportedInBatchMode
@@ -583,44 +429,6 @@ respond stateVar request =
         case Map.lookup mainName state.definitions of
             Nothing -> pure $ Left $ RpcError.backendError $ RpcError.CouldNotFindModule mainName
             Just d -> action (d, state.mLlvmLibrary, state.mSMTOptions, state.rewriteOptions)
-
-    doesNotImply' s condition l r =
-        pure $
-            Right $
-                RpcTypes.Implies
-                    RpcTypes.ImpliesResult
-                        { implication = addHeader $ Syntax.KJImplies s l r
-                        , valid = False
-                        , condition
-                        , logs = Nothing
-                        }
-
-    doesNotImply s' = let s = externaliseSort s' in doesNotImply' s Nothing
-    implies' predicate s l r subst =
-        pure $
-            Right $
-                RpcTypes.Implies
-                    RpcTypes.ImpliesResult
-                        { implication = addHeader $ Syntax.KJImplies s l r
-                        , valid = True
-                        , condition =
-                            Just
-                                RpcTypes.Condition
-                                    { predicate = addHeader predicate
-                                    , substitution =
-                                        addHeader
-                                            $ ( \case
-                                                    [] -> Syntax.KJTop s
-                                                    [x] -> x
-                                                    xs -> Syntax.KJAnd s xs
-                                              )
-                                                . map (uncurry $ externaliseSubstitution s)
-                                                . Map.toList
-                                            $ subst
-                                    }
-                        , logs = Nothing
-                        }
-    implies s' = let s = externaliseSort s' in implies' (Syntax.KJTop s) s
 
 handleSmtError :: JsonRpcHandler
 handleSmtError = JsonRpcHandler $ \case
